pyproject.toml
[project]
name = "agent-prompt-dkb"
version = "0.1.0"
description = "Agent prompt curation instance powered by DKB"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "dkb-runtime",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["agent_prompt_dkb"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
